pyproject.toml
[project]
name = "skn23-3rd-2team"
version = "0.1.0"
description = "현장 기술지원 챗봇 (Robot/Welding)"
authors = [
    {name = "JUYEOP024", email = "juyeobsong63@gmail.com"}
]
readme = "README.md"
requires-python = "~3.12"

dependencies = [
    "langchain (==1.2.10)",
    "langchain-openai (==1.1.10)",
    "langchain-community (==0.4.1)",
    "langgraph (==1.0.9)",
    "chromadb (==1.5.1)",
    "langchain-chroma (==1.1.0)",
    "fastapi (==0.129.0)",
    "uvicorn (==0.41.0)",
    "pydantic (==2.12.5)",
    "pydantic-settings (==2.13.1)",
    "pymysql (==1.1.2)",
    "bcrypt (==5.0.0)",
    "streamlit (==1.54.0)",
    "requests (==2.32.5)",
    "python-dotenv (==1.2.1)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"